clean up code structure
Mild refactoring, and also make pyright happy because it makes me feel better
